refactor: Separate threeJS object creation into Node subclasses. [#20]
