+++ index.html
@@ -6,7 +6,7 @@
     <title>Inochi2D example</title>
   </head>
   <body>
-    <div id="app"></div>
+    <canvas id="app"></canvas>
     <script type="module" src="/lib/test/test.ts">
 
     </script>
