+++ lib/nodes/node.ts
@@ -5,9 +5,9 @@
     Authors: Luna Nielsen
 */
 
-import { Transform, deserializeTransform } from "../math/transform";
-import { MeshData } from "../meshdata";
-import { Vector2 } from "three";
+import { Transform } from "../math/transform";
+import * as THREE from "three";
+import { blend_modes } from "../renderer/renderer";
 
 /**
  * Blending mode.
@@ -55,6 +55,14 @@ export enum MaskingMode {
 }
 
 /**
+ * Represents the joint binding data.
+ */
+export class JointBindingData {
+    bound_to: NodeUuid = -1;
+    bind_data: number[][] = [];
+}
+
+/**
  * Represents the Inox Node UUID.
  */
 export type NodeUuid = number;
@@ -63,6 +71,7 @@ export type NodeUuid = number;
  * Base type for all nodes.
  */
 export class Node {
+    // Serialised models
     type: string = "";
     uuid: NodeUuid = -1;
     name?: string;
@@ -71,216 +80,79 @@ export class Node {
     transform: Transform = new Transform();
     children: Node[] = [];
 
-    // Private non-serialisables
-    parent: Node | null = null;
-    lockToRoot: boolean = false;
-    actualTransform: Transform = new Transform();
-    actualZsort: number = 0;
+    // Non-serialisables
+    threeObj: THREE.Object3D | THREE.Mesh = new THREE.Object3D();   // three.JS Node for rendering.
+    parent: Node | null = null;                                     // Track the parent for easy traversal.
+    lockToRoot: boolean = false;                                    // Whether to lock to root
+    actualTransform: Transform = new Transform();                   // Track absolute transform  
+    actualZsort: number = 0;                                        // Track absolute z-index
 
     /**
      * Calculates the transform of this node.
      */
     updateTransform() {
-        // Update transforms
+        // Update transform
         this.transform.update();
 
-        // If this has a parent, orient transform in relation to parent
+        // Update the absolute transform
         if (this.parent == null) {
+            // No parent, use current transform
             this.actualTransform = this.transform;
             this.actualZsort = this.zsort;
         } else {
+            // Otherwise, translate, rotate & scale current transform to the parent's
             const newTransform = new Transform();
             newTransform.rot = this.parent.actualTransform.rot.clone().add(this.transform.rot);
             newTransform.trans = this.parent.actualTransform.trans.clone().add(this.transform.trans);
-            newTransform.scale = new Vector2(this.parent.actualTransform.scale.x * this.transform.scale.x,
+            newTransform.scale = new THREE.Vector2(this.parent.actualTransform.scale.x * this.transform.scale.x,
                 this.parent.actualTransform.scale.y * this.transform.scale.y);
 
+            // Set the transform
             this.actualTransform = newTransform;
             this.actualZsort = this.parent.actualZsort + this.zsort;
         }
+
+        // Update the three object transform
+        this.threeObj.position.set(this.transform.trans.x, this.transform.trans.y, this.transform.trans.z);  
+        this.threeObj.scale.set(this.transform.scale.x, this.transform.scale.y, 1); 
+        this.threeObj.rotation.x = this.transform.rot.x;
+        this.threeObj.rotation.y = this.transform.rot.y;
+        this.threeObj.rotation.z = this.transform.rot.z;
+        this.threeObj.renderOrder = -this.actualZsort;
     }
 
     /**
-     * Cycles a node.
+     * Called on render, populates a THREE.Object3D.
+     */
+    protected onCreateMesh() {}
+
+    /**
+     * Called on render, populates a THREE.Object3D materials.
+     */
+    protected onCreateMaterials(textures: THREE.Texture[]) {}
+
+    /**
+     * Called on update.
      */
     update() {
         this.updateTransform();
     }
+
+    /**
+     * Called on render.
+     */
+    create(textures: THREE.Texture[]) {
+        this.onCreateMesh();
+        this.onCreateMaterials(textures);
+    }
 }
 
-/**
- * Represents the drawable properties.
- */
-export class Drawable extends Node {
-    mesh: MeshData = new MeshData();
-}
-
-/**
- * Represents a part with additional properties.
- */
-export class Part extends Drawable {
-    textures: number[] = [];
-    opacity: number = 1;
-    mask_mode: MaskingMode = MaskingMode.Mask;
-    mask_threshold: number = 0;
-    masked_by: NodeUuid[] = [];
-    blend_mode: BlendMode = BlendMode.Normal;
-}
-
-/**
- * Represents a mask with the same properties as a drawable.
- */
-export class Mask extends Drawable { }
 
 /**
  * Represents a path deform node.
  */
 export class PathDeform extends Node {
-    joints: Vector2[] = [];
+    joints: THREE.Vector2[] = [];
     bindings: JointBindingData[] = [];
 }
 
-/**
- * Represents the joint binding data.
- */
-export class JointBindingData {
-    bound_to: NodeUuid = -1;
-    bind_data: number[][] = [];
-}
-
-function deserializeBaseProperties(json: any, node: Node): void {
-    node.type = json.type;
-    node.uuid = json.uuid;
-    node.name = json.name;
-    node.enabled = json.enabled !== undefined ? json.enabled : node.enabled;
-    node.zsort = json.zsort !== undefined ? json.zsort : node.zsort;
-    node.transform = json.transform !== undefined ? deserializeTransform(json.transform) : node.transform;
-    node.children = json.children !== undefined ? json.children.map((child: any) => deserializeNode(child, node)) : node.children;
-    node.lockToRoot = json.children !== undefined ? json.lockToRoot : node.lockToRoot;
-}
-
-function deserializeDrawable(json: any, drawable: Drawable): Drawable {
-    deserializeBaseProperties(json, drawable);
-    // Deserialize additional properties specific to Drawable
-    drawable.mesh = MeshData.deserialize(json.mesh);
-    return drawable;
-}
-
-function deserializePart(json: any): Part {
-    const part = new Part();
-    deserializeDrawable(json, part);
-    // Deserialize additional properties specific to Part
-    part.textures = json.textures;
-    part.opacity = json.opacity;
-    part.mask_mode = json.mask_mode;
-    part.mask_threshold = json.mask_threshold;
-    part.masked_by = json.masked_by;
-
-    if (json.blend_mode)
-        switch (json.blend_mode) {
-            case "Multiply":
-                part.blend_mode = BlendMode.Multiply;
-                break;
-            case "ColorDodge":
-                part.blend_mode = BlendMode.ColorDodge;
-                break;
-            case "LinearDodge":
-                part.blend_mode = BlendMode.LinearDodge;
-                break;
-            case "Screen":
-                part.blend_mode = BlendMode.Screen;
-                break;
-            default:
-                part.blend_mode = BlendMode.Normal;
-        }
-
-    return part;
-}
-
-function deserializeMask(json: any): Mask {
-    const mask = new Mask();
-    deserializeDrawable(json, mask);
-    // Deserialize additional properties specific to Mask
-    return mask;
-}
-
-function deserializePathDeform(json: any): PathDeform {
-    const pathDeform = new PathDeform();
-    deserializeBaseProperties(json, pathDeform);
-    // Deserialize additional properties specific to PathDeform
-    pathDeform.joints = json.joints;
-    pathDeform.bindings = json.bindings;
-    return pathDeform;
-}
-
-function deserializeCustomNode(json: any): Node {
-    const node = new Node();
-    deserializeBaseProperties(json, node);
-    return node;
-}
-
-/**
- * Deserializes a JSON object into the appropriate subclass based on the "type" property.
- * @param json - The JSON object to deserialize.
- * @param parent - The Node that may have called this.
- * @returns The deserialized object as the correct subclass.
- */
-export function deserializeNode(json: any, parent: Node | null = null): Node {
-    let result = new Node();
-    switch (json.type) {
-        case "Part":
-            result = deserializePart(json);
-            break;
-        case "Mask":
-            result = deserializeMask(json);
-            break;
-        case "PathDeform":
-            result = deserializePathDeform(json);
-            break;
-        default:
-            result = deserializeCustomNode(json);
-            break;
-    }
-
-    // Set the parent
-    result.parent = parent;
-
-    return result;
-}
-
-/**
- * Serializes a Node object into a JSON object with the specified properties.
- * @param node - The Node object to serialize.
- * @returns The serialized JSON object.
- */
-export function serializeNode(node: Node): any {
-    const data: any = {};
-
-    data.type = node.type;
-    data.uuid = node.uuid;
-    data.name = node.name;
-    data.enabled = node.enabled;
-    data.zsort = node.zsort;
-    data.transform = node.transform;
-    data.children = node.children.map(child => serializeNode(child,));
-    data.lockToRoot = node.lockToRoot;
-
-    if (node instanceof Drawable) {
-        data.mesh = node.mesh;
-    }
-
-    if (node instanceof Part) {
-        data.textures = node.textures;
-        data.opacity = node.opacity;
-        data.mask_mode = node.mask_mode;
-        data.mask_threshold = node.mask_threshold;
-        data.masked_by = node.masked_by;
-    }
-
-    if (node instanceof PathDeform) {
-        data.joints = node.joints;
-        data.bindings = node.bindings;
-    }
-
-    return data;
-}
