+++ tsconfig.json
@@ -10,11 +10,11 @@
     "isolatedModules": true,
     "esModuleInterop": true,
     "noEmit": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
     "noImplicitReturns": true,
     "skipLibCheck": true,
-    "allowJs": true
+    "allowJs": true,
+    "noUnusedLocals": false,
+    "noUnusedParameters": false
   },
   "include": ["src", "lib"]
 }
